Refactored entity extensions
Each entity gets its own file to prevent a long single file from being used
