+++ src/Tsa.Submissions.Coding.WebApi/Entities/EntityExtensions.Participant.cs
@@ -1,0 +1,22 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using Tsa.Submissions.Coding.WebApi.Models;
+
+namespace Tsa.Submissions.Coding.WebApi.Entities;
+
+public static partial class EntityExtensions
+{
+    public static ParticipantModel ToModel(this Participant participant)
+    {
+        return new ParticipantModel
+        {
+            ParticipantNumber = participant.ParticipantNumber,
+            SchoolNumber = participant.SchoolNumber
+        };
+    }
+
+    public static List<ParticipantModel> ToModels(this IList<Participant> participants)
+    {
+        return participants.Select(participant => participant.ToModel()).ToList();
+    }
+}
