+++ src/Tsa.Submissions.Coding.WebApi/Entities/EntityExtensions.Problem.cs
@@ -1,0 +1,24 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using Tsa.Submissions.Coding.WebApi.Models;
+
+namespace Tsa.Submissions.Coding.WebApi.Entities;
+
+public static partial class EntityExtensions
+{
+    public static ProblemModel ToModel(this Problem problem)
+    {
+        return new ProblemModel
+        {
+            Description = problem.Description,
+            Id = problem.Id,
+            IsActive = problem.IsActive,
+            Title = problem.Title
+        };
+    }
+
+    public static List<ProblemModel> ToModels(this IList<Problem> problems)
+    {
+        return problems.Select(problem => problem.ToModel()).ToList();
+    }
+}
